Detect a file type ELF
Some distribution packages have a link to the 'run' file to an
executable file type ELF.

For example, the 'run' file of the 'log' service:

<service>/log/run -> /usr/bin/vlogger

The 'xsv' file editor obviously cannot edit these files.
The user is informed of what is happening and writing in the
editor is not allowed.